Update Makefile for image registry and name; add 'name' field to CRDs and API documentation
This commit modifies the Makefile to change the image registry and name for the Vitess operator. Additionally, it introduces a new 'name' field in several CRDs (etcdlockservers, vitesscells, vitessclusters, vitesskeyspaces, vitessshards) to allow for optional service name overrides. The API documentation is also updated to reflect these changes, enhancing clarity on the usage of the new field.